# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CoupleRequest": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "from_user": {
            "$ref": "#/components/schemas/UserRef"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "to_user": {
            "$ref": "#/components/schemas/UserRef"
          }
        },
        "required": [
          "id",
          "from_user",
          "to_user",
          "created_at"
        ],
        "title": "CoupleRequest",
        "type": "object"
      },
      "CoupleRequestCreate": {
        "properties": {
          "user_id": {
            "maxLength": 36,
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "CoupleRequestCreate",
        "type": "object"
      },
      "CoupleState": {
        "description": "The own couple (`partner`, accepted at `since`) and pending requests.",
        "properties": {
          "incoming": {
            "items": {
              "$ref": "#/components/schemas/CoupleRequest"
            },
            "title": "Incoming",
            "type": "array"
          },
          "outgoing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CoupleRequest"
              },
              {
                "type": "null"
              }
            ]
          },
          "partner": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "since": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Since"
          }
        },
        "required": [
          "partner",
          "since",
          "outgoing",
          "incoming"
        ],
        "title": "CoupleState",
        "type": "object"
      },
...
      "ErrorCode": {
        "description": "What went wrong; the frontend shows the translation `error.<code>`.",
        "enum": [
          "common.internal",
          "common.not_found",
          "common.method_not_allowed",
          "common.validation",
          "common.rate_limited",
          "common.unauthorized",
          "common.forbidden",
          "common.service_unavailable",
          "auth.invalid_credentials",
          "auth.account_deactivated",
          "auth.token_expired",
          "auth.session_expired",
          "auth.session_revoked",
          "auth.login_required",
          "auth.csrf",
          "auth.code_invalid",
          "auth.password_incorrect",
          "couple.already_in_couple",
          "couple.target_in_couple",
          "couple.request_pending",
          "admin.self_forbidden",
          "admin.last_admin",
          "admin.public_url_missing"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "description": "`params` fill placeholders in the translation; `fields` lists rejected request fields.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Fields",
            "type": "array"
          },
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code",
          "params",
          "fields"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FieldError": {
        "description": "One rejected request field; `loc` is its path, e.g. `[\"body\", \"name\"]`.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/FieldErrorCode"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          }
        },
        "required": [
          "loc",
          "code"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "FieldErrorCode": {
        "description": "Why a single request field was rejected (`fields[].code` in the envelope).",
        "enum": [
          "required",
          "invalid",
          "too_short",
          "too_long",
          "out_of_range",
          "invalid_format",
          "taken",
          "too_common",
          "same_as_username"
        ],
        "title": "FieldErrorCode",
        "type": "string"
      },
      "FilterHidden": {
        "description": "User filter chips this user has switched off, per screen (MEAL-10, UI-02).",
        "properties": {
          "lists": {
            "items": {
              "maxLength": 36,
              "type": "string"
            },
            "maxItems": 500,
            "title": "Lists",
            "type": "array"
          },
          "meals": {
            "items": {
              "maxLength": 36,
              "type": "string"
            },
            "maxItems": 500,
            "title": "Meals",
            "type": "array"
          }
        },
        "required": [
          "meals",
          "lists"
        ],
        "title": "FilterHidden",
        "type": "object"
      },
...
      "Me": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "filter_hidden": {
            "$ref": "#/components/schemas/FilterHidden"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "language": {
            "enum": [
              "de",
              "en"
            ],
            "title": "Language",
            "type": "string"
          },
          "lists_public": {
            "title": "Lists Public",
            "type": "boolean"
          },
          "meals_public": {
            "title": "Meals Public",
            "type": "boolean"
          },
          "role": {
            "enum": [
              "user",
              "admin"
            ],
            "title": "Role",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "display_name",
          "role",
          "language",
          "meals_public",
          "lists_public",
          "filter_hidden",
          "created_at"
        ],
        "title": "Me",
        "type": "object"
      },
      "MeUpdate": {
        "description": "Only the fields that are sent change (ACC-13, VIS-02, I18N-01).",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "filter_hidden": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FilterHidden"
              },
              {
                "type": "null"
              }
            ]
          },
          "language": {
            "anyOf": [
              {
                "enum": [
                  "de",
                  "en"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "lists_public": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lists Public"
          },
          "meals_public": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meals Public"
          }
        },
        "title": "MeUpdate",
        "type": "object"
      },
      "PasswordChange": {
        "properties": {
          "current_password": {
            "maxLength": 1024,
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "maxLength": 1024,
            "title": "New Password",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChange",
        "type": "object"
      },
...
      "SecurityInfo": {
        "description": "What *Me → Security* shows, e.g. \"Password reset by X on <date>\" (ACC-10).\n\n`password_reset_by` is null for a reset link from the command line or a deleted admin.",
        "properties": {
          "password_changed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password Changed At"
          },
          "password_reset_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password Reset At"
          },
          "password_reset_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "password_changed_at",
          "password_reset_at",
          "password_reset_by"
        ],
        "title": "SecurityInfo",
        "type": "object"
      },
      "SessionInfo": {
        "description": "A logged-in device; `current` is the one making this request (ACC-09).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current": {
            "title": "Current",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_used_at": {
            "format": "date-time",
            "title": "Last Used At",
            "type": "string"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          }
        },
        "required": [
          "id",
          "user_agent",
          "created_at",
          "last_used_at",
          "current"
        ],
        "title": "SessionInfo",
        "type": "object"
      },
      "UserRef": {
        "description": "Another user as everyone sees them; deactivated users are shown with a note (ADM-02).",
        "properties": {
          "deactivated": {
            "title": "Deactivated",
            "type": "boolean"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "deactivated"
        ],
        "title": "UserRef",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "The access token from login, refresh, join or reset.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "license": {
      "identifier": "AGPL-3.0-or-later",
      "name": "AGPL-3.0-or-later"
    },
    "title": "MealMate",
    "version": "0.0.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/couple": {
      "delete": {
        "description": "End the couple (either partner, any time).",
        "operationId": "end_couple",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "End Couple",
        "tags": [
          "couple"
        ]
      },
      "get": {
        "description": "The partner, if any, and pending requests.",
        "operationId": "get_couple",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoupleState"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Couple",
        "tags": [
          "couple"
        ]
      }
    },
    "/api/couple/requests": {
      "post": {
        "description": "Send a couple request to another active user.",
        "operationId": "create_couple_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoupleRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoupleState"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Couple Request",
        "tags": [
          "couple"
        ]
      }
    },
    "/api/couple/requests/{request_id}/accept": {
      "post": {
        "description": "Accept a request sent to you; cancels all other pending requests of both users.",
        "operationId": "accept_couple_request",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoupleState"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Accept Couple Request",
        "tags": [
          "couple"
        ]
      }
    },
    "/api/couple/requests/{request_id}/cancel": {
      "post": {
        "description": "Withdraw your own request.",
        "operationId": "cancel_couple_request",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoupleState"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Couple Request",
        "tags": [
          "couple"
        ]
      }
    },
    "/api/couple/requests/{request_id}/decline": {
      "post": {
        "description": "Decline a request sent to you.",
        "operationId": "decline_couple_request",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoupleState"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Decline Couple Request",
        "tags": [
          "couple"
        ]
      }
    },
...
    "/api/me": {
      "get": {
        "description": "The own profile and settings.",
        "operationId": "get_me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "me"
        ]
      },
      "patch": {
        "description": "Change display name, language, privacy switches or hidden filter chips.",
        "operationId": "update_me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Me",
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/password": {
      "post": {
        "description": "Change the password; logs out all other devices.",
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Password",
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/security": {
      "get": {
        "description": "Password change and reset notices (\"Password reset by X on <date>\").",
        "operationId": "get_security",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Security",
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/sessions": {
      "get": {
        "description": "The own logged-in devices, most recently used first.",
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionInfo"
                  },
                  "title": "Response List Sessions",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "me"
        ]
      }
    },
    "/api/me/sessions/{session_id}": {
      "delete": {
        "description": "Log out one of the own devices.",
        "operationId": "revoke_session",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Session",
        "tags": [
          "me"
        ]
      }
    },
...
  }
}